refactor: rule 패키지 이름 game으로 변경
