+++ src/main/java/blackjack/controller/BlackjackGame.java
@@ -5,7 +5,7 @@ import blackjack.domain.participant.Dealer;
 import blackjack.domain.participant.Participants;
 import blackjack.domain.participant.Player;
 import blackjack.domain.participant.Players;
-import blackjack.domain.rule.Judge;
+import blackjack.domain.game.Judge;
 import blackjack.view.InputView;
 import blackjack.view.MessageResolver;
 import blackjack.view.OutputView;
