+++ src/main/java/blackjack/domain/game/Result.java
@@ -1,4 +1,4 @@
-package blackjack.domain;
+package blackjack.domain.game;
 
 import java.util.List;
