+++ src/test/java/blackjack/domain/card/HandTest.java
@@ -2,7 +2,7 @@ package blackjack.domain.card;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import blackjack.domain.rule.Score;
+import blackjack.domain.game.Score;
 import java.util.ArrayList;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
