+++ src/test/java/blackjack/domain/game/JudgeTest.java
@@ -1,6 +1,5 @@
-package blackjack.domain.rule;
+package blackjack.domain.game;
 
-import blackjack.domain.Result;
 import blackjack.domain.card.Card;
 import blackjack.domain.card.CardRank;
 import blackjack.domain.card.CardSuit;
