+++ src/test/java/blackjack/domain/game/ResultTest.java
@@ -1,8 +1,9 @@
-package blackjack.domain;
+package blackjack.domain.game;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.*;
 
+import blackjack.domain.game.Result;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
